requirements.txt
aiogram==3.10.0
telethon==1.36.0
python-dotenv==1.0.1
APScheduler==3.10.4
pytest==7.4.4
pytest-asyncio==0.23.6
